# docs/openapi/smartload-v1.yaml
openapi: 3.1.0
info:
  title: SmartLoad API
  version: 1.0.0
  summary: Operating-policy, telemetry, scaling, and audit API for the SmartLoad stack.
  description: |
    Canonical contract for the SmartLoad HTTP surface. Every shipped
    `/api/v1/*` route appears here. Generated client SDKs (`smartload-client`
    for Python) and the operator UI both consume this spec.

    Stability: every breaking change requires a path-version bump
    (`/api/v2`) and a `Sunset` header on the deprecated version. See SOT
    §11.5 (versioning policy).
servers:
  - url: http://localhost:8086
    description: policy-manager (local dev)
  - url: http://localhost:8081
    description: telemetry (local dev)
...
paths:
...
  /api/v1/audit/scaling:
    get:
      summary: Recent autoscaler-action audit rows
      description: |
        One row per scaling action taken by the autoscaler (scale-out or
        scale-in), newest first. Read-only — the autoscaler writes
        scaling_events when it actuates; this endpoint replays them for
        operators and integrators.

        Lives on the autoscaler service (default port 8085), not on
        policy-manager.
      tags: [autoscaler, audit]
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 50
      responses:
        '200':
          description: Recent scaling-event rows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScalingAuditRow'
        '400':
          description: Invalid limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Audit datastore unreachable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/scale:
    post:
      summary: Manually scale the backend pool
      description: |
        Operator override on the autoscaler (port 8085). Bypasses the
        forecast subscription and the cooldown timer. Validates the
        requested `target_count` against the live policy's
        `min_backends` / `max_backends` and rejects 400 when out of
        bounds. On success, scales the cluster step-by-step, writes one
        `scaling_events` row with `reason = "manual:<actor>: <reason>"`,
        and publishes one `ScalingEvent` envelope on `smartload.scale`
        (with `forecast_event_id = null` because there is no forecast
        origin).
      tags: [autoscaler]
      parameters:
        - in: header
          name: X-Actor
          required: false
          schema: { type: string }
          description: |
            Logical actor persisted in the audit row reason. If unset,
            the request body `actor` field is used; if both are unset,
            defaults to `"operator"`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManualScaleRequest'
      responses:
        '200':
          description: Scale applied (or no-op if target equals current)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManualScaleResponse'
        '400':
          description: Target out of bounds or non-integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Cluster query or audit write failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    ScalingAuditRow:
      type: object
      required: [time, action, instance_count]
      properties:
        time:           { type: string, format: date-time }
        action:
          type: string
          enum: [scale_out, scale_in]
          description: Direction of the scaling action.
        instance_count:
          type: integer
          description: Resulting backend count after the action.
        reason:
          type: string
          nullable: true
          description: Human-readable justification (e.g. "forecast predicted 450 rps").
    ManualScaleRequest:
      type: object
      required: [target_count]
      properties:
        target_count:
          type: integer
          description: |
            Desired backend count after the action. Must be in
            `[policy.min_backends, policy.max_backends]`.
        actor:
          type: string
          description: |
            Logical actor name. Falls back to the `X-Actor` header, then
            to `"operator"`.
        reason:
          type: string
          description: |
            Free-text justification. Prefixed with `manual:<actor>:` in
            the audited row.
    ManualScaleResponse:
      type: object
      required: [status, action, target_count, previous_count, final_count, steps_actuated, steps_requested, reason, event_id]
      properties:
        status:
          type: string
          enum: [applied, noop]
        action:
          type: string
          enum: [scale_out, scale_in, noop]
        target_count:    { type: integer }
        previous_count:  { type: integer }
        final_count:     { type: integer }
        steps_actuated:  { type: integer }
        steps_requested: { type: integer }
        reason:          { type: string }
        event_id:        { type: string, format: uuid }
...
    Error:
      type: object
      required: [error]
      properties:
        error:   { type: string }
        field:
          type: [string, "null"]
          description: |
            Names the offending field when the server can pinpoint it
            (validation errors). Null otherwise.

# Out of v1.0.0 (filed but not yet specified):
#   POST /api/v1/scale       (#123)
#   POST /api/v1/isolate     (#123)
#   POST /api/v1/webhooks    (#130)
#   GET  /api/v1/webhooks    (#130)
#   DELETE /api/v1/webhooks/{id}  (#130)
#   POST /api/v1/api-keys    (#132)